Introduce `newclient` package to break the dependency cycle when moving towards v3
