+++ pkg/releases/create_release_v1.go
@@ -1,4 +1,4 @@
-package executionsapi
+package releases
 
 import (
 	"encoding/json"
