+++ test/e2e/release_service_test.go
@@ -1,7 +1,6 @@
 package e2e
 
 import (
-	"github.com/OctopusDeploy/go-octopusdeploy/v2/pkg/executionsapi"
 	"testing"
 
 	"github.com/OctopusDeploy/go-octopusdeploy/v2/internal"
@@ -159,8 +158,8 @@ func TestReleaseServiceCreateV1(t *testing.T) {
 	require.NotNil(t, channel)
 	defer DeleteTestChannel(t, client, channel)
 
-	releaseCreate := executionsapi.NewCreateReleaseCommandV1(space.Name, project.Name)
-	createReleaseResponse, err := executionsapi.CreateReleaseV1(client, releaseCreate)
+	releaseCreate := releases.NewCreateReleaseCommandV1(space.Name, project.Name)
+	createReleaseResponse, err := releases.CreateReleaseV1(client, releaseCreate)
 
 	// if HTTP 404 response then Executions API is unavailable
 	if err == nil {
